feat: incapsulate class `BasicStructureGenerator` in separate file, import it into `index.js`.00
