+++ src/index.js
@@ -1,8 +1,6 @@
 // ESLint rules per file start
 // ESLint rules per file end
 
-// import '../pages/index.html';
-
 /**
  * Utility's import.
  */
@@ -15,14 +13,18 @@
 //   addAllFieldsAsTextNode,
 // } from "./utils";
 
-
 /**
  * App phrase content import.
  */
 import {
   /* Objects and vars: */
   phrase,
-} from "./content";
+} from './content';
+
+/**
+ * Basic layout generator import.
+ */
+import BasicStructureGenerator from './BasicStructureGenerator';
 
 /* Import Pics */
 // Example: import birdPlaceholder from '../assets/images/raven_01.png';
@@ -41,27 +43,7 @@ const debugFlag = 0;
 const debug = debugHardcode ? debugHardcode : debugFlag;
 
 /* Constants and vars of states and other things. */
-let languageGeneral = "en";
-
-class BasicStructureGenerator {
-  root;
-
-  rootToken;
-
-  constructor(rootToken) {
-    this.rootToken = rootToken;
-    this.root = document.querySelector(rootToken);
-  }
-
-  generate() {
-    const testElement = document.createElement('div');
-    testElement.style.width = '256px';
-    testElement.style.height = '256px';
-    testElement.style.backgroundColor = 'blue';
-
-    this.root.appendChild(testElement);
-  }
-}
+let languageGeneral = 'en';
 
 const basicStructureGenerator = new BasicStructureGenerator('body');
